+++ packages/sanity-jsonschema-forms/src/index.ts
@@ -1,12 +1,10 @@
-export type {SupportedFieldType} from './to-json-schema'
+export type {SupportedFieldType} from './internal/field'
+export {FORM_TOOLKIT_FIELD_TYPES, isSupportedType, SUPPORTED_FIELD_TYPES} from './internal/field'
 export {
   COLOR_PATTERN,
   DATETIME_LOCAL_PATTERN,
-  FORM_TOOLKIT_FIELD_TYPES,
-  isSupportedType,
   JSON_SCHEMA_DRAFT_7,
   NONZERO_YEAR_PATTERN,
-  SUPPORTED_FIELD_TYPES,
   TIME_PATTERN,
   toJsonSchema,
 } from './to-json-schema'